Update Player class to use float values for position coordinates

+++ src/characters/player.py
@@ -76,7 +76,7 @@ class Player(Person):
 
     # Properties
     @property
-    def pos(self) -> tuple[int, int]:
+    def pos(self) -> tuple[float, float]:
         """Get the position of the player.
 
         Returns:
